+++ src/Models/Attribute.php
@@ -10,9 +10,9 @@ class Attribute extends BaseAttribute
     /**
      * Build the validation rules for the attribute field type.
      */
-    public function fieldTypeValidations(): array
+    public function fieldTypeValidations(?int $id = null, array $allowedPathPrefixes = []): array
     {
-        $rules = parent::fieldTypeValidations();
+        $rules = parent::fieldTypeValidations($id, $allowedPathPrefixes);
 
         switch ($this->type) {
             case 'asset':