+++ src/Resources/views/components/layouts/with-history/asset.blade.php
@@ -83,10 +83,11 @@
                 <x-admin::layouts.sidebar />
                 
                                 
-                <div class="flex-1 max-w-full overflow-y-auto overflow-x-hidden px-4 pt-3 pb-6 bg-transparent dark:bg-cherry-800 ltr:pl-[286px] rtl:pr-[286px] max-lg:!px-4 transition-all duration-300 group-[.sidebar-collapsed]/container:ltr:pl-[85px] group-[.sidebar-collapsed]/container:rtl:pr-[85px]">
+                <div class="flex-1 min-w-0 overflow-y-auto overflow-x-hidden px-4 pt-3 pb-6 bg-transparent dark:bg-cherry-800 max-lg:!px-4 transition-all duration-300">
                     {!! view_render_event('unopim.admin.layouts.tabs.before') !!}
 
-                    
+                    <x-admin::breadcrumbs :leaf="$label ?? ''" class="mb-3" />
+
                     <div class="flex flex-wrap justify-between gap-2 items-center">
                         <div class="flex min-w-0">