+++ vite.config.js
@@ -1,6 +1,6 @@
 import { defineConfig, loadEnv } from "vite";
 import vue from "@vitejs/plugin-vue";
-import laravel from 'laravel-vite-plugin';
+import laravel from "laravel-vite-plugin";
 import path from "path";
 
 export default defineConfig(({ mode }) => {
@@ -11,6 +11,7 @@ export default defineConfig(({ mode }) => {
     return {
         build: {
             emptyOutDir: true,
+            minify: "esbuild",
         },
 
         envDir,
@@ -24,7 +25,7 @@ export default defineConfig(({ mode }) => {
             vue(),
             
             laravel({
-                hotFile: "../../../public/dam.hot",
+                hotFile: "../../../public/dam-vite.hot",
                 publicDirectory: "../../../public",
                 buildDirectory: "themes/dam/build",
                 input: [